+++ CHANGELOG.md
@@ -5,9 +5,10 @@ Please add your entries according to this format.
 
 ## Unreleased
 
+- Remove dropboxStyle since it is no longer supported by ktfmt. Use kotlinLangStyle() instead
 - Fix task caching for ktfmtCheckTask when project has multiple source sets (#288)
 - Kotlin to 2.0.0
-- ksp to 2.0.0-1.0.22
+- KtFmt to 0.51
 
 ## Version 0.18.0 _(2024-04-13)_